almost finished

+++ sh/replaySingularity.sh
@@ -14,32 +14,32 @@ ml GCC/9.3.0 Singularity/3.7.3-Go-1.14
 
 # Passed parameters
 PARAM=("${@}")
-INPUT_DIR=${PARAM[0]}
-RUN_DIR=${PARAM[2]}
-SPLIT=${PARAM[4]}
-OUTPUT_DIR=${PARAM[11]}
+INPUT_DIR=${PARAM[1]}
+RUN_DIR=${PARAM[3]}
+SPLIT=${PARAM[6]}
+OUTPUT_DIR=${PARAM[4]}
 
 # Is this a regular job or a job array
 if [[ -n $SLURM_ARRAY_TASK_ID ]]
 then
   JOB_ID=$SLURM_ARRAY_TASK_ID
+  # Get the ID so in the R script we can know either the region index or the travel time
   PARAM+=("$JOB_ID") 
   # New directories for binding data folder (bypass denied access for writing) and out directory
+  # We need a database per job to avoid conflict (in the end of this script we remove these folders)
   mkdir -p "$INPUT_DIR/AMdata/temp/$JOB_ID/dbgrass"
   mkdir -p "$INPUT_DIR/AMdata/temp/$JOB_ID/cache"
   mkdir -p "$INPUT_DIR/AMdata/temp/$JOB_ID/logs"
-  # Files to be binded
+  # Directory to be binded
   DATA_DIR="$INPUT_DIR/AMdata/temp/$JOB_ID"
 else
   mkdir -p "$INPUT_DIR/AMdata/dbgrass"
   mkdir -p "$INPUT_DIR/AMdata/cache"
   mkdir -p "$INPUT_DIR/AMdata/logs"
   DATA_DIR="$INPUT_DIR/AMdata"
-fi
-
-if [[ $SPLIT == "true" ]]
-then
-  REGION_JSON_FILE=${PARAM[14]}
+  JOB_ID=""
+  # So we we have the same number of parameters that are passed to the R script
+  PARAM+=("$JOB_ID")
 fi
 
 # Other dir/files to be binded
@@ -47,21 +47,37 @@ PROJECT_FILE="$INPUT_DIR/project.am5p"
 R_SCRIPT_FILE="$RUN_DIR/R/replay.R"
 CONFIG_FILE="$INPUT_DIR/config.json"
 
-# Input
-
-echo "Start processing AccessMod Job"
-
-# Run image with binded inputs and launch R script
-singularity run \
+# If split we need to bind the regions.json file as well
+if [[ $SPLIT == "true" ]]
+then
+  REGION_JSON_FILE=${PARAM[14]}
+  # Run image with binded inputs and launch R script
+  echo "Start processing AccessMod Job"
+  singularity run \
+    -B $OUTPUT_DIR:/batch/out \
+    -B $PROJECT_FILE:/batch/project.am5p \
+    -B $CONFIG_FILE:/batch/config.json \
+    -B $REGION_JSON_FILE:/batch/regions.json \
+    -B $R_SCRIPT_FILE:/batch/replay.R \
+    -B $DATA_DIR:/data \
+    --pwd /app \
+    $IMAGE \
+    Rscript /batch/replay.R "${PARAM[@]}"
+else
+  echo "Start processing AccessMod Job"
+  singularity run \
   -B $OUTPUT_DIR:/batch/out \
   -B $PROJECT_FILE:/batch/project.am5p \
   -B $CONFIG_FILE:/batch/config.json \
-  -B $INPUT_FILE:/batch/inputs.json \
   -B $R_SCRIPT_FILE:/batch/replay.R \
   -B $DATA_DIR:/data \
   --pwd /app \
   $IMAGE \
   Rscript /batch/replay.R "${PARAM[@]}"
+fi
 
 # Remove duplicated data (dbgrass, logs, cache)
-rm -r "$INPUT_DIR/AMdata/temp/"
+if [[ -n $SLURM_ARRAY_TASK_ID ]]
+then
+  rm -r "$INPUT_DIR/AMdata/temp/"
+fi
